+++ app/models/user.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Integer, String, Uuid, func
+from sqlalchemy import BIGINT, Boolean, DateTime, String, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.common.models import Base
@@ -12,7 +12,7 @@ class User(Base):
 
     uid: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
     user_name: Mapped[str] = mapped_column(String(255), nullable=False, unique=True)
-    user_tg_id: Mapped[int] = mapped_column(Integer, nullable=False, unique=True)
+    user_tg_id: Mapped[int] = mapped_column(BIGINT, nullable=False, unique=True)
     is_bot: Mapped[bool] = mapped_column(Boolean, default=False)
     firstname: Mapped[str | None] = mapped_column(String(255), nullable=True)
     lastname: Mapped[str | None] = mapped_column(String(255), nullable=True)
